+++ bloom_nofos/nofos/nofo.py
@@ -545,31 +545,31 @@ def join_nested_lists(soup):
                 return ps
         return None
 
-    def _join_lists(ul, previous_ul):
-        if _get_list_classname(ul) == _get_list_classname(previous_ul):
+    def _join_lists(lst, previous_lst):
+        if _get_list_classname(lst) == _get_list_classname(previous_lst):
             # if classes match, join these lists
-            previous_ul.extend(ul.find_all("li"))
-            ul.decompose()
+            previous_lst.extend(lst.find_all("li"))
+            lst.decompose()
             return
 
         # okay: classes do not match
         # get the last li in the previous list
-        last_tag_in_previous_list = previous_ul.find_all("li", recursive=False)[-1]
+        last_tag_in_previous_list = previous_lst.find_all("li", recursive=False)[-1]
 
-        # see if there is a ul in there
-        nested_ul = last_tag_in_previous_list.find("ul")
-        if nested_ul:
-            return _join_lists(ul, nested_ul)
+        # see if there is a ul/ol in there
+        nested_lst = last_tag_in_previous_list.find(["ul", "ol"])
+        if nested_lst:
+            return _join_lists(lst, nested_lst)
 
         # if there is not, append to the last li
-        last_tag_in_previous_list.append(ul)
+        last_tag_in_previous_list.append(lst)
         return
 
-    for ul in soup.find_all("ul"):
+    for lst in soup.find_all(["ul", "ol"]):
         # check previous sibling
-        previous_element = _get_previous_element(ul)
-        if previous_element and previous_element.name == "ul":
-            _join_lists(ul, previous_element)
+        previous_element = _get_previous_element(lst)
+        if previous_element and previous_element.name in ["ul", "ol"]:
+            _join_lists(lst, previous_element)
 
     return soup