+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 ### Added
 
 - Added section ids to the table captions on hover in edit mode
+- Nested lists now work with ols as well as uls
+  - Supports intermixing them as well (eg: ul, ol, ul)
 
 ### Changed